US-06 backend

+++ backend-java/PXL-Nieuwsartikelen/post-service/src/main/java/be/pxl/services/domain/dto/PostRequest.java
@@ -19,7 +19,7 @@ public class PostRequest {
     private String title;
     private String content;
     private String author;
-    private int authorId;
+    private Long authorId;
     private String category;
     private State state;
     private LocalDateTime createdAt;
